Use nullglob to prevent PATH errors
When setting the list of path directories, we sometimes use bash
globbing to account for variances in versions. For example, to add the
binaries for PostgreSQL 9.4 to the path, we could add:

    /usr/lib/postgresql/9.4/bin

to the path_dirs variable. However, when postgresql updates, this will
break. Instead, we can use a glob to account for all versions:

    /usr/lib/postgresql/*/bin

On systems in which postgresql is not installed, this will throw an
error. Using bash's nullglob option prevents this by expanding failed
globs to null strings.